Move keyword and nets parser to parsers subdirectory
Move the keyword and nets parser to the parsers subdirectory.

* Add `src/parsers/keywords.rs` with `VerilogKeyword` enum, `keyword_from_string` function, and tests.
* Add `src/parsers/nets.rs` with `NetType` enum, `net_type` function, and tests.
* Delete `src/keywords.rs`.
* Modify `src/utils.rs` to remove `NetType` enum and `net_type` function.
* Modify `src/main.rs` to update imports to use `parsers::keywords` and `parsers::nets`.
